+++ src/test/java/roomescape/reservation/unit/repository/FakeReservationRepository.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.repository;
+package roomescape.reservation.unit.repository;
 
 import java.util.ArrayList;
 import java.util.Collections;
