+++ src/test/java/roomescape/reservationtime/unit/domain/ReservationTimeTest.java
@@ -1,4 +1,4 @@
-package roomescape.reservationtime.domain;
+package roomescape.reservationtime.unit.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
