+++ src/test/java/roomescape/reservationtime/unit/service/ReservationTimeServiceTest.java
@@ -1,4 +1,4 @@
-package roomescape.reservationtime.service;
+package roomescape.reservationtime.unit.service;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
@@ -10,8 +10,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import roomescape.reservationtime.domain.ReservationTime;
 import roomescape.reservationtime.dto.ReservationTimeRequest;
-import roomescape.reservationtime.repository.FakeReservationTimeRepository;
 import roomescape.reservationtime.repository.ReservationTimeRepository;
+import roomescape.reservationtime.unit.repository.FakeReservationTimeRepository;
 
 class ReservationTimeServiceTest {
